+++ src/Nix/Thunk/Command.hs
@@ -54,6 +54,7 @@ data ThunkCommand
   | ThunkCommand_Unpack (NonEmpty FilePath)
   | ThunkCommand_Worktree CreateWorktreeConfig (FilePath, FilePath)
   | ThunkCommand_Pack ThunkPackConfig (NonEmpty FilePath)
+  | ThunkCommand_Stage ThunkConfig (NonEmpty FilePath)
   | ThunkCommand_Create ThunkCreateConfig
   deriving Show
 
@@ -76,6 +77,7 @@ thunkCommand = hsubparser $ mconcat
   , command "unpack" $ info (ThunkCommand_Unpack <$> thunkDirList) $ progDesc "Unpack thunk into git checkout of revision it points to"
   , command "worktree" $ info (ThunkCommand_Worktree <$> createWorktreeConfig <*> createWorktreeArgs) $ progDesc "Create a git worktree of the thunk using the specified local git repo"
   , command "pack" $ info (ThunkCommand_Pack <$> thunkPackConfig <*> thunkDirList) $ progDesc "Pack git checkout or unpacked thunk into thunk that points at the current branch's upstream"
+  , command "stage" $ info (ThunkCommand_Stage <$> thunkConfig <*> thunkDirList) $ progDesc "Stage a packed representation of an unpacked thunk without changing its checkout"
   , command "create" $ info (ThunkCommand_Create <$> thunkCreateConfig) $ progDesc "Create a packed thunk without cloning the repository first"
   ]
 
@@ -93,4 +95,5 @@ runThunkCommand = \case
   ThunkCommand_Unpack dirs -> mapM_ unpackThunk dirs
   ThunkCommand_Worktree config (thunkDir, gitDir) -> createWorktree thunkDir gitDir config
   ThunkCommand_Pack config dirs -> mapM_ (packThunk config) dirs
+  ThunkCommand_Stage config dirs -> mapM_ (stageThunk config) dirs
   ThunkCommand_Create config -> createThunk' config